+++ src/main/java/telran/currency/service/CurrencyConvertor.java
@@ -6,6 +6,6 @@ import java.util.List;
 public interface CurrencyConvertor {
    List<String> strongestCurrencies(int amount);
    List<String> weakestCurrencies(int amount);
-   double convert(String codeFrom, String codeTo, int amount);
+   double convert(String codeFrom, String codeTo, double amount);
    HashSet<String> getAllCodes();
 }
